restructured script code, added MIT license to used Sortable code
