+++ pipelinediagrams.html
@@ -7,7 +7,7 @@
       href="https://fonts.googleapis.com/css?family=Montserrat:400,400i,700"
     />
     <link rel="stylesheet" href="styles.css" />
-    <script type="module" src="script.js"></script>
+    <script type="module" src="scripts/setup.js"></script>
     <title>RISC-V Processor</title>
   </head>
   <body>
@@ -218,7 +218,7 @@
     </div>
     </div>
     <footer>
-      <p>Copyright &#169; 2024 <a href="/">My Site</a></p>
+      <p>Copyright &#169; 2024 <a href="https://github.com/Maremas/RISCVEDU">Tim Gaisbauer, Marius Henrich</a></p>
     </footer>
   </body>
 </html>
